Remove outdated configmap tests and instructions; update pod annotation instructions
