--- tests/game01/24_Create_a_configmap/answer.template.yaml
@@ -1,7 +1,0 @@
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: config
-data:
-  foo: lala
-  foo2: lolo
