--- tests/game01/24_Create_a_configmap/instruction.md
@@ -1,1 +1,0 @@
-Create a configmap named config with values foo=lala,foo2=lolo
