--- tests/game01/24_Create_a_configmap/test_01_setup.py
@@ -1,6 +1,0 @@
-import pytest
-
-
-@pytest.mark.order(1)
-def test_setup(json_input):
-    pass
