--- tests/game01/24_Create_a_configmap/test_06_cleanup.py
@@ -1,26 +1,0 @@
-# test_06_cleanup.py
-# test_06_cleanup.py
-import logging
-import pytest
-from tests.helper.kubectrl_helper import build_kube_config, run_kubectl_command
-
-class TestCleanup:
-
-    @pytest.mark.order(6)
-    def test_cleanup(self, json_input):
-        # 配置 Kubernetes
-        kube_config = build_kube_config(
-            json_input["cert_file"], json_input["key_file"], json_input["host"]
-        )
-        
-        namespace = json_input["namespace"]
-        
-        # 删除 ConfigMap 的命令
-        configmap_name = "config"
-        command_configmap = f"kubectl delete configmap {configmap_name} -n {namespace}"
-        
-        # 运行命令并记录结果
-        result_configmap = run_kubectl_command(kube_config, command_configmap)
-        
-        logging.info(result_configmap)
-        assert "deleted" in result_configmap.lower(), f"Failed to delete ConfigMap '{configmap_name}' in namespace '{namespace}'"
