+++ tests/game01/24_create_a_configmap/answer.template.yaml
@@ -1,0 +1,13 @@
+apiVersion: v1
+kind: Namespace
+metadata:
+  name: {{namespace}}
+---
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: config
+  namespace: {{namespace}}
+data:
+  foo: lala
+  foo2: lolo
