+++ tests/game01/24_create_a_configmap/session.json
