+++ tests/game01/24_create_a_configmap/setup.template.yaml
@@ -1,0 +1,4 @@
+apiVersion: v1
+kind: Namespace
+metadata:
+  name: {{namespace}}
