+++ tests/game01/24_create_a_configmap/test_01_setup.py
@@ -1,0 +1,8 @@
+import pytest
+
+from tests.helper.test_helper import deploy_setup
+
+
+@pytest.mark.order(1)
+def test_setup(json_input):
+    deploy_setup(json_input)
