+++ tests/game01/24_create_a_configmap/test_03_answer.py
@@ -1,0 +1,8 @@
+import pytest
+
+from tests.helper.test_helper import deploy_answer
+
+
+@pytest.mark.order(3)
+def test_answer(json_input):
+    deploy_answer(json_input)
