+++ tests/game01/24_create_a_configmap/test_05_check.py
@@ -1,9 +1,12 @@
 # test_05_check_configmap.py
-import logging
-import pytest
 import json
+import logging
+
+import pytest
+
 from tests.helper.k8s_client_helper import configure_k8s_client
 from tests.helper.kubectrl_helper import build_kube_config, run_kubectl_command
+
 
 @pytest.mark.order(5)
 class TestCheckConfigMap:
@@ -13,9 +16,11 @@ class TestCheckConfigMap:
         # 获取ConfigMap
         configmap_name = "config"
         namespace = json_input["namespace"]
-        
+
         try:
-            configmap = k8s_client.read_namespaced_config_map(name=configmap_name, namespace=namespace)
+            configmap = k8s_client.read_namespaced_config_map(
+                name=configmap_name, namespace=namespace
+            )
         except Exception as e:
             assert False, f"Failed to get ConfigMap '{configmap_name}': {str(e)}"
 
@@ -27,21 +32,21 @@ class TestCheckConfigMap:
         assert configmap.data["foo"] == "lala", "Incorrect value for 'foo'."
         assert "foo2" in configmap.data, "Missing key 'foo2' in data."
         assert configmap.data["foo2"] == "lolo", "Incorrect value for 'foo2'."
-        logging.info(f"ConfigMap '{configmap_name}' has the correct content.")
+        logging.info("ConfigMap '%s' has the correct content.", configmap_name)
 
     def test_002_check_configmap_kubectl(self, json_input):
         logging.debug("Starting test_002_check_configmap_kubectl")
         kube_config = build_kube_config(
             json_input["cert_file"], json_input["key_file"], json_input["host"]
         )
-        
+
         # 使用kubectl获取ConfigMap
         command = f"kubectl get configmap config -n {json_input['namespace']} -o json"
         result = run_kubectl_command(kube_config, command)
-        
+
         # 解析kubectl输出的JSON内容
         configmap = json.loads(result)
-        
+
         # 验证ConfigMap的内容
         assert configmap["apiVersion"] == "v1", "Incorrect apiVersion."
         assert configmap["kind"] == "ConfigMap", "Incorrect kind."
@@ -50,8 +55,4 @@ class TestCheckConfigMap:
         assert configmap["data"]["foo"] == "lala", "Incorrect value for 'foo'."
         assert "foo2" in configmap["data"], "Missing key 'foo2' in data."
         assert configmap["data"]["foo2"] == "lolo", "Incorrect value for 'foo2'."
-        logging.info(f"ConfigMap 'config' has the correct content.")
-
-# 运行测试
-if __name__ == "__main__":
-    pytest.main()
+        logging.info("ConfigMap 'config' has the correct content.")
